Castling is now possible and the AI recognizes it as a move
git-svn-id: https://svn.neumont.edu:8443/svn/jpearl_chess@44 86f9d0df-9005-f54b-96e6-c011f00df492

+++ src/edu/neumont/pro180/jpearl/chess/pieces/Piece.java
@@ -16,9 +16,9 @@ public abstract class Piece
 	private int numberOfMoves;
 	private int lastTurn;
 	private char characterRepresentation;
-	private int unitWorth;
+	private double unitWorth;
 	
- 	public Piece( PieceColor color, char characterRepresentation, int unitWorth )
+ 	public Piece( PieceColor color, char characterRepresentation, double unitWorth )
 	{
  		this.characterRepresentation = characterRepresentation;
  		this.unitWorth = unitWorth;
@@ -61,7 +61,7 @@ public abstract class Piece
 		return characterRepresentation;
 	}
 	
-	public int getUnitWorth()
+	public double getUnitWorth()
 	{
 		return unitWorth;
 	}
